requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
black==22.10.0
cffi==1.15.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-model-utils==4.2.0
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
flake8-black==0.3.3
gunicorn==20.1.0
#hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
#psycopg2==2.9.3
#psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.5.0
pyOpenSSL==22.1.0
python-decouple==3.6
pytz==2022.1
service-identity==21.1.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
whitenoise==6.0.0
zope.interface==5.5.0
requests==2.28.2
